Update exchange-config to use $STATIC_ROOT
In the django function there is a "python manage.py collectstatic"
followed by a chown and a chmod, previously if a user was using
a custom static directory then the files would be moved to the
custom directory and the default directory would get chown/mod-ed,
leaving the working static directory untouched. This change insures
that the static directory will always be updated appropriately,
regardless of the use of standard/custom directories.