Add variable to toggle Extent Filter on Search Page
The extent filter uses a leaflet tile map in order
to filter results based on the contents shown
in the map. This map has to reach out across the
open internet in order to properly access the
tiles.

For some deploys this is not ideal, as it is
not possible to access the open internet due
to firewalls or other restrictions. As such it
was determined that simply disabling the filter
was the best way forward for those types of
networks.